# api/pkg/interface/gen/openapi/server.go
//go:build go1.22

// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package openapi

import (
	"errors"
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for CreateTodoObjectiveStatus.
const (
	CreateTodoObjectiveStatusArchive    CreateTodoObjectiveStatus = "archive"
	CreateTodoObjectiveStatusCompleted  CreateTodoObjectiveStatus = "completed"
	CreateTodoObjectiveStatusInProgress CreateTodoObjectiveStatus = "in_progress"
	CreateTodoObjectiveStatusNotStarted CreateTodoObjectiveStatus = "not_started"
)

// Valid indicates whether the value is a known member of the CreateTodoObjectiveStatus enum.
func (e CreateTodoObjectiveStatus) Valid() bool {
	switch e {
	case CreateTodoObjectiveStatusArchive:
		return true
	case CreateTodoObjectiveStatusCompleted:
		return true
	case CreateTodoObjectiveStatusInProgress:
		return true
	case CreateTodoObjectiveStatusNotStarted:
		return true
	default:
		return false
	}
}

// Defines values for TodoStatus.
const (
	TodoStatusArchive    TodoStatus = "archive"
	TodoStatusCompleted  TodoStatus = "completed"
	TodoStatusInProgress TodoStatus = "in_progress"
	TodoStatusNotStarted TodoStatus = "not_started"
)

// Valid indicates whether the value is a known member of the TodoStatus enum.
func (e TodoStatus) Valid() bool {
	switch e {
	case TodoStatusArchive:
		return true
	case TodoStatusCompleted:
		return true
	case TodoStatusInProgress:
		return true
	case TodoStatusNotStarted:
		return true
	default:
		return false
	}
}

// Defines values for UpdateTodoObjectiveStatus.
const (
	UpdateTodoObjectiveStatusArchive    UpdateTodoObjectiveStatus = "archive"
	UpdateTodoObjectiveStatusCompleted  UpdateTodoObjectiveStatus = "completed"
	UpdateTodoObjectiveStatusInProgress UpdateTodoObjectiveStatus = "in_progress"
	UpdateTodoObjectiveStatusNotStarted UpdateTodoObjectiveStatus = "not_started"
)

// Valid indicates whether the value is a known member of the UpdateTodoObjectiveStatus enum.
func (e UpdateTodoObjectiveStatus) Valid() bool {
	switch e {
	case UpdateTodoObjectiveStatusArchive:
		return true
	case UpdateTodoObjectiveStatusCompleted:
		return true
	case UpdateTodoObjectiveStatusInProgress:
		return true
	case UpdateTodoObjectiveStatusNotStarted:
		return true
	default:
		return false
	}
}

// CreateTodoObjective defines model for CreateTodoObjective.
type CreateTodoObjective struct {
	Status CreateTodoObjectiveStatus `json:"status"`
	Title  string                    `json:"title"`
}

// CreateTodoObjectiveStatus defines model for CreateTodoObjective.Status.
type CreateTodoObjectiveStatus string

// Error defines model for Error.
type Error struct {
	Code       string `json:"code"`
	Message    string `json:"message"`
	StatusCode int    `json:"status_code"`
}

// Pagination defines model for Pagination.
type Pagination struct {
	PageNumber uint `json:"page_number"`
	PageSize   int  `json:"page_size"`
	Total      uint `json:"total"`
}

// Todo defines model for Todo.
type Todo struct {
	Id     openapi_types.UUID `json:"id"`
	Status TodoStatus         `json:"status"`
	Title  string             `json:"title"`
}

// TodoStatus defines model for Todo.Status.
type TodoStatus string

// UpdateTodoObjective defines model for UpdateTodoObjective.
type UpdateTodoObjective struct {
	Status UpdateTodoObjectiveStatus `json:"status"`
	Title  string                    `json:"title"`
}

// UpdateTodoObjectiveStatus defines model for UpdateTodoObjective.Status.
type UpdateTodoObjectiveStatus string

// ID defines model for ID.
type ID = openapi_types.UUID

// PageNumber defines model for PageNumber.
type PageNumber = uint

// PageSize defines model for PageSize.
type PageSize = int

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Created defines model for Created.
type Created = Todo

// InternalServerError defines model for InternalServerError.
type InternalServerError = Error

// NotFound defines model for NotFound.
type NotFound = Error

// TodoObjective defines model for TodoObjective.
type TodoObjective = Todo

// TodoWithPagination defines model for TodoWithPagination.
type TodoWithPagination struct {
	Data       []Todo `json:"data"`
	PageNumber uint   `json:"page_number"`
	PageSize   int    `json:"page_size"`
	Total      uint   `json:"total"`
}

// GetTodoListParams defines parameters for GetTodoList.
type GetTodoListParams struct {
	PageNumber *PageNumber `form:"page_number,omitempty" json:"page_number,omitempty"`

	// PageSize If you want to fetch all records, specify a negative value.
	PageSize *PageSize `form:"page_size,omitempty" json:"page_size,omitempty"`
}

// CreateTodoObjectiveJSONRequestBody defines body for CreateTodoObjective for application/json ContentType.
type CreateTodoObjectiveJSONRequestBody = CreateTodoObjective

// UpdateTodoObjectiveJSONRequestBody defines body for UpdateTodoObjective for application/json ContentType.
type UpdateTodoObjectiveJSONRequestBody = UpdateTodoObjective

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetHealth Health Application check
	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)
	// GetTodoList get Todo list
	// (GET /todos)
	GetTodoList(w http.ResponseWriter, r *http.Request, params GetTodoListParams)
	// CreateTodoObjective create Todo Objective
	// (POST /todos)
	CreateTodoObjective(w http.ResponseWriter, r *http.Request)
	// DeleteTodoObjective delete Todo Objective
	// (DELETE /todos/{id})
	DeleteTodoObjective(w http.ResponseWriter, r *http.Request, id ID)
	// GetTodoByID get Todo By ID
	// (GET /todos/{id})
	GetTodoByID(w http.ResponseWriter, r *http.Request, id ID)
	// UpdateTodoObjective update Todo Objective
	// (PATCH /todos/{id})
	UpdateTodoObjective(w http.ResponseWriter, r *http.Request, id ID)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTodoList operation middleware
func (siw *ServerInterfaceWrapper) GetTodoList(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTodoListParams

	// ------------- Optional query parameter "page_number" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page_number", r.URL.Query(), &params.PageNumber, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "page_number"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page_number", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "page_size" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page_size", r.URL.Query(), &params.PageSize, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "page_size"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page_size", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTodoList(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateTodoObjective operation middleware
func (siw *ServerInterfaceWrapper) CreateTodoObjective(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateTodoObjective(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteTodoObjective operation middleware
func (siw *ServerInterfaceWrapper) DeleteTodoObjective(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteTodoObjective(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTodoByID operation middleware
func (siw *ServerInterfaceWrapper) GetTodoByID(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTodoByID(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateTodoObjective operation middleware
func (siw *ServerInterfaceWrapper) UpdateTodoObjective(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateTodoObjective(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/health", wrapper.GetHealth)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/todos", wrapper.GetTodoList)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/todos", wrapper.CreateTodoObjective)
	m.HandleFunc(http.MethodDelete+" "+options.BaseURL+"/todos/{id}", wrapper.DeleteTodoObjective)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/todos/{id}", wrapper.GetTodoByID)
	m.HandleFunc(http.MethodPatch+" "+options.BaseURL+"/todos/{id}", wrapper.UpdateTodoObjective)

	return m
}